Relayout of the JavaScript client library to ESM, part 2
The main MatsSocket.js library is now a EcmaScript Modules library, with
package.json having type="module", with Rollup bundling to UMD files.

Had to give in on the ".js" ending for the bundled
"MatsSocket.umd[.min].js" files (now ending in ".cjs"), as otherwise
node insists on treating it as a ES module when including it from the
tests-subproject (because of its containing package.json has that
type="module"), which I assume also would hold if getting it via
published npm.