Relayout of the JavaScript client library to ESM, part 4
Now the "HTML tests", i.e. unit and integration tests running in the
browser, served from MatsSocketTestServer, exists in two variants:

1. ESM where everything is EcmaScript Modules based, both the tests, and
the MatsSocket.js.

2. UMD, which uses the built file 'MatsSocket.umd.csj', as well as still
using the old copies of the tests, with UMD preamble.

If it isn't clear, both of these sets of tests are run when doing
./gradlew distClean check test build, and it is the exact same files
that are served in the browser. This was actually a bit problematic, as
node uses "bare" imports, e.g. "import * as mats from 'matssocket'",
while the browser expects URLs in the reference. So, other people have
also realized this, and Import maps is the answer. Implemented by
Chrome, while Firefox seemingly first doesn't see the light, and now
just drags their feet.