Removed not-under-test ACK'ed check from test.
Got a fail on GHA MacOS - and it turns out that the MatsSocketServer's
async handling of "non-information-bearing-messages" (ACK, NACK, ACK2
etc) could easily lead to this situation. I.e. that the receivedCallback
("ACK") can come later than the server-initiated SEND to the terminator.

Point is that in this test, the SEND from client-to-server is totally
disjoint from the SEND from server-to-client (other than the first
triggering the latter). Compare this to a a REQUEST: If the REPLY comes
in earlier than the ACK, the client just "inserts" an ACK handling in
front of the REPLY handling, guaranteeing the ordering.