Choosing to delete from repo, and .gitignore 'package-lock.json'.
We delete it upon 'distclean'.
All deps are only for building the lib. They are resolved and downloaded
based on package.json.

This then aligns with the Dart-client's non-commit of 'pubspec.lock',
which is customary to NOT include for libraries in Dart-world, even
though it evidently is customary to do commit 'package-lock.json' in JS
world. Breakin' the law.