Oops, had made the JS unit tests not work in browser anymore
The import didn't take in browser, as Mocha is not loaded as an ESM
module - those 'describe()' etc are globals. Running them cli in Node,
mocha itself adds them as globals. So, only place needed is in IDE, to
not get red squiggles when JSHint and ESlint is enabled. This change got
those gone in IntelliJ.