Adding ClusterStoreAndForward_SQL.useSetNString(.) and .getNString(.) to tailor DB interaction.
For MS SQL, it is important that one do NOT use
'sendStringParametersAsUnicode=true' to fix all encoding problems, as
this nukes performance. Thus, if employing the older-skool NVARCHAR for
long-text column types, it is better to set
'sendStringParametersAsUnicode=false', and then specifically target the
NVARCHAR columns with stmt.setNString(idx, string).

This feature makes that possible.